@@ -20,17 +20,17 @@
     "tslint": "tslint -c tslint.json \"src/**/*.ts\" -e src/app/datamodels/attachment-data-types.ts --force"
   },
   "dependencies": {
-    "@angular/common": "^2.1.2",
-    "@angular/compiler": "^2.1.2",
-    "@angular/compiler-cli": "^2.1.2",
-    "@angular/core": "^2.1.2",
-    "@angular/forms": "^2.1.2",
-    "@angular/http": "^2.1.2",
-    "@angular/platform-browser": "^2.1.2",
-    "@angular/platform-browser-dynamic": "^2.1.2",
-    "@angular/platform-server": "^2.1.2",
-    "@angular/router": "^3.1.2",
-    "@angular/upgrade": "^2.1.2",
+    "@angular/common": "^2.2.1",
+    "@angular/compiler": "^2.2.1",
+    "@angular/compiler-cli": "^2.2.1",
+    "@angular/core": "^2.2.1",
+    "@angular/forms": "^2.2.1",
+    "@angular/http": "^2.2.1",
+    "@angular/platform-browser": "^2.2.1",
+    "@angular/platform-browser-dynamic": "^2.2.1",
+    "@angular/platform-server": "^2.2.1",
+    "@angular/router": "^3.2.1",
+    "@angular/upgrade": "^2.2.1",
     "@angularclass/hmr": "^1.2.2",
     "@angularclass/hmr-loader": "^3.0.2",
     "@angularclass/request-idle-callback": "^1.0.7",
@@ -46,33 +46,33 @@
     "material-design-icons": "^3.0.1",
     "material-icons": "^0.1.0",
     "material-icons-font": "^1.0.0",
-    "ng2-translate": "^4.0.0",
+    "ng2-translate": "^4.0.1",
     "ngrx-store-logger": "^0.1.7",
     "node-libs-browser": "^1.0.0",
     "reflect-metadata": "^0.1.8",
-    "rxjs": "^5.0.0-rc.2",
+    "rxjs": "^5.0.0-rc.4",
     "zone.js": "^0.6.25"
   },
   "devDependencies": {
-    "@ngtools/webpack": "^1.1.4",
+    "@ngtools/webpack": "^1.1.7",
     "@types/chai": "^3.4.34",
     "@types/chrome": "0.0.35",
     "@types/core-js": "^0.9.34",
     "@types/hammerjs": "^2.0.33",
-    "@types/jasmine": "^2.5.35",
+    "@types/jasmine": "^2.5.38",
     "@types/jquery": "^2.0.33",
-    "@types/node": "^6.0.45",
+    "@types/node": "^6.0.48",
     "@types/selenium-webdriver": "^2.53.33",
     "@types/source-map": "^0.1.28",
     "@types/uglify-js": "^2.6.28",
-    "@types/webpack": "^1.12.35",
-    "angular-cli": "^1.0.0-beta.18",
+    "@types/webpack": "^1.12.36",
+    "angular-cli": "^1.0.0-beta.20-4",
     "angular2-template-loader": "^0.5.0",
     "awesome-typescript-loader": "^2.2.4",
     "babel": "^6.5.2",
     "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-class-properties": "^6.18.0",
+    "babel-loader": "^6.2.8",
+    "babel-plugin-transform-class-properties": "^6.19.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.14.0",
     "clean-webpack-plugin": "^0.1.10",
@@ -88,13 +88,14 @@
     "html-minify-loader": "^1.1.0",
     "html-webpack-plugin": "^2.22.0",
     "intl": "^1.2.5",
+    "istanbul-instrumenter-loader": "~0.2.0",
     "jasmine-core": "^2.5.2",
     "karma": "^1.3.0",
     "karma-chrome-launcher": "^2.0.0",
     "karma-coverage": "^1.1.1",
     "karma-env-preprocessor": "^0.1.1",
     "karma-jasmine": "^1.0.2",
-    "karma-mocha-reporter": "^2.2.0",
+    "karma-mocha-reporter": "^2.2.1",
     "karma-phantomjs-launcher": "^1.0.2",
     "karma-remap-coverage": "^0.1.2",
     "karma-sourcemap-loader": "^0.3.7",
@@ -121,10 +122,10 @@
     "tslint": "^3.15.1",
     "tslint-loader": "^2.1.5",
     "typedoc": "^0.5.0",
-    "typescript": "^2.0.9",
+    "typescript": "~2.0.9",
     "uglify-loader": "^1.3.0",
     "url-loader": "^0.5.7",
-    "webpack": "^2.1.0-beta.22",
+    "webpack": "^2.1.0-beta.27",
     "webpack-dev-middleware": "^1.8.4",
     "webpack-dev-server": "^2.1.0-beta.2",
     "webpack-merge": "^0.15.0"
