@@ -20,8 +20,7 @@ const METADATA = webpackMerge(commonConfig({env: ENV}).metadata, {});
 const helpers = require('./helpers');
 
 const webpackOptions = {
-    filesToIgnore: ["icon_19.png"],
-    defaultIcon: "snapchat.png",
+    filesToCopy: [{ from: '../src/icons_dev', to: "./icons", toType: "dir", flatten: true }],
     env: ENV,
     cleanOutput: false
 };
