@@ -1,7 +1,1 @@
-declare namespace twemoji {
-    function parse(what: string, how?: any): string;
-}
-
-declare namespace twemoji.convert {
-    function fromCodePoint(codepoint: string): string;
-}
+export declare const twemoji;
