@@ -1,5 +1,5 @@
 /*! Copyright Twitter Inc. and other contributors. Licensed under MIT */
-var twemoji = function() {
+export var twemoji = function() {
   "use strict";
   var twemoji = {
       base: (location.protocol === "https:" ? "https:" : "http:") + "//abs.twimg.com/emoji/v2/",
