@@ -13,7 +13,7 @@ export class FileUploadService {
     constructor(private vkservice: VKService) { }
 
     getServerUrl(): Observable<string> {
-        return this.vkservice.performAPIRequest(this.getServerMethod, null)
+        return this.vkservice.performAPIRequestsBatch(this.getServerMethod, null)
             .map(response => {
                 console.log("server url", response);
                 return response.upload_url;
@@ -21,7 +21,7 @@ export class FileUploadService {
     }
 
     getPhoto(photo: Photo): Observable<string> {
-        return this.vkservice.performAPIRequest(this.savePhotoMethod, `photo=${photo.photo}&server=${photo.server}&hash=${photo.hash}`)
+        return this.vkservice.performAPIRequestsBatch(this.savePhotoMethod, {photo: photo.photo, server: photo.server, hash: photo.hash})
             .map(response => {
                 console.log("got photo to send", response);
                 return `photo${response[0].owner_id}_${response[0].id}`;
