@@ -1,5 +1,5 @@
-///<reference path="../../../lib/twemoji.d.ts"/>
 import { Pipe } from "@angular/core";
+import { twemoji } from "../../../lib/twemoji";
 
 @Pipe({ name: "emoji" })
 export class EmojiPipe {
