@@ -3,17 +3,18 @@ import { Observable } from "rxjs/Observable";
 import { Store } from "@ngrx/store";
 
 import { VKService } from "./vk.service";
-import { SingleMessageInfo, DialogInfo, DialogListInfo, ChatInfo, DialogShortInfo, HistoryInfo } from "../datamodels";
+import { SingleMessageInfo, DialogInfo, DialogListInfo, ChatInfo, DialogShortInfo, HistoryInfo, InputMessageInfo } from "../datamodels";
 import { Channels } from "../../app.background/channels";
 import { ChromeAPIService } from "./chrome-api.service";
-import { AppStore, HistoryActions } from "../app.store";
+import { AppState, HistoryActions } from "../app.store";
+import { typeMessage } from "../actions";
 
 @Injectable()
 export class DialogService {
     constructor(
         private vkservice: VKService,
         private chromeapi: ChromeAPIService,
-        private store: Store<AppStore>
+        private store: Store<AppState>
     ) { }
 
     init() {
@@ -28,15 +29,26 @@ export class DialogService {
         }).map(x => x.data);
     }
 
-    sendMessage(id: number, message: any, chat: boolean): Observable<number> {
+    sendMessage(id: number, message: any, chat: boolean): void {
         console.log("sending message", message);
-        return this.chromeapi.SendRequest({
-            name: Channels.sendMessageRequest,
-            user_id: id,
-            message_body: message.body,
-            attachments: message.attachments,
-            is_chat: chat
-        }).map(x => x.data);
+        let inputMessageInfo = {
+            conversationId: id,
+            chatId: (chat ? id : null),
+            body: message.body,
+            attachments: message.attachments
+        } as InputMessageInfo;
+        this.chromeapi.PostPortMessage({ name: "send_message", data: inputMessageInfo });
+    }
+
+    typeMessage(id: number, message: any, chat: boolean): void {
+        console.log("type message", id, message, chat);
+        let inputMessageInfo = {
+            conversationId: id,
+            chatId: (chat ? id : null),
+            body: message.body,
+            attachments: message.attachments
+        } as InputMessageInfo;
+        this.chromeapi.PostPortMessage({ name: "type_message", data: inputMessageInfo });
     }
 
     searchDialog(searchTerm: string): Observable<DialogShortInfo[]> {
